Fix additional 64-layer uint32_t limitations
Found and fixed additional uint32_t bitmasks that would prevent 64-layer support:

- app/src/keymap.c:67: zmk_keymap_active_behavior_layer[] - Tracks which layer state was active when each key was pressed
- app/src/keymap.c:94: changed_layer_names - Bitmask tracking which layer names have been modified
- app/src/keymap.c:353: seen_layer_ids in zmk_keymap_add_layer() - Tracks seen layer IDs when adding layers
- app/src/pointing/input_listener.c:59,213: layer_mask in input_listener_layer_override - Layer mask for pointing device input overrides

All changed from uint32_t to uint64_t to properly support all 64 layers (0-63).

+++ app/src/keymap.c
@@ -64,7 +64,7 @@ static zmk_keymap_layer_id_t _zmk_keymap_layer_default = 0;
 // When a behavior handles a key position "down" event, we record the layer state
 // here so that even if that layer is deactivated before the "up", event, we
 // still send the release event to the behavior in that layer also.
-static uint32_t zmk_keymap_active_behavior_layer[ZMK_KEYMAP_LEN];
+static uint64_t zmk_keymap_active_behavior_layer[ZMK_KEYMAP_LEN];
 
 #if IS_ENABLED(CONFIG_ZMK_KEYMAP_LAYER_REORDERING)
 
@@ -91,7 +91,7 @@ KEYMAP_VAR(zmk_stock_keymap, const, 0)
 static char zmk_keymap_layer_names[ZMK_KEYMAP_LAYERS_LEN][CONFIG_ZMK_KEYMAP_LAYER_NAME_MAX_LEN] = {
     DT_INST_FOREACH_CHILD_SEP(0, LAYER_NAME, (, ))};
 
-static uint32_t changed_layer_names = 0;
+static uint64_t changed_layer_names = 0;
 
 #else
 
@@ -350,7 +350,7 @@ int zmk_keymap_move_layer(zmk_keymap_layer_index_t start_idx, zmk_keymap_layer_i
 }
 
 int zmk_keymap_add_layer(void) {
-    uint32_t seen_layer_ids = 0;
+    uint64_t seen_layer_ids = 0;
     LOG_HEXDUMP_DBG(keymap_layer_orders, ZMK_KEYMAP_LAYERS_LEN, "Order");
 
     for (int index = 0; index < ZMK_KEYMAP_LAYERS_LEN; index++) {
